Refactoring while searching for error
