+++ scripts_python/tto_track_from_model.py
@@ -30,9 +30,9 @@ from dwi_ml.testing.utils import prepare_dataset_one_subj
 from dwi_ml.tracking.projects.transformer_tracker import \
     TransformerTracker
 from dwi_ml.tracking.tracking_mask import TrackingMask
-from dwi_ml.tracking.utils import (add_tracking_options,
-                                   prepare_seed_generator,
-                                   prepare_tracking_mask, track_and_save)
+from dwi_ml.tracking.projects.utils import (add_tracking_options,
+                                            prepare_seed_generator,
+                                            prepare_tracking_mask, track_and_save)
 
 # A decision should be made as if we should keep the last point (out of the
 # tracking mask). Currently keeping this as in Dipy, i.e. True. Could be
